Migrate Mutation.createOIDCIntegration
